[ALPS05804815] FPSGO: cm_tdiff cannot work
[Detail]
Cast variables to type long long before add operation

MTK-Commit-Id: b0765acfaee9c60e5d1ff024c2ba6988aaa7b5bb

CR-Id: ALPS05804815
Feature: FPSGO

+++ drivers/misc/mediatek/performance/fpsgo_v3/fbt/src/fbt_cpu.c
@@ -2057,7 +2057,8 @@ static void fbt_check_cm_limit(struct render_info *thread_info,
 		return;
 
 	if (last_blc > cm_big_cap &&
-		runtime > thread_info->boost_info.target_time + cm_tdiff)
+		runtime > (long long)thread_info->boost_info.target_time +
+		(long long)cm_tdiff)
 		fbt_notify_CM_limit(1);
 	else
 		fbt_notify_CM_limit(0);
